(search) move filter-box, filter-option and select to components and update search-publishers-tree
