+++ applications/search/src/components/filter-box/filter-box.component.test.jsx
