+++ applications/search/src/components/filter-option/filter-option.component.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { FilterOption } from './filter-option.component';
-import filterOptions from '../../../../test/fixtures/filterOptions';
+import filterOptions from '../../../test/fixtures/filterOptions';
 
 test('should render FilterOption correctly', () => {
   const onClick = jest.fn();
