+++ applications/search/src/components/select/select.component.test.jsx
